requirements.txt
requests
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
pytest-redis
jsonpath
PyYAML